requirements.txt

djangorestframework==3.10.2
djangorestframework-yaml==1.0.3
